Issue #8: Refactor SankeySettingsPanel component -- all settings moved to dedicated submodules.
